DataProcessing/requirements.txt
numpy
scipy
cython
h5py
pandas
PyMySQL
seaborn
scikit-learn
tensorflow-gpu==1.15.2
cryptography